- Core 2.0 -> 3.0 마이그레이션 - 닷넷 커뮤니티 사이트로 변경 - 기타 보안 처리 - 기타 여러가지 오류들 수정

+++ RVCoreBoard.MVC/Apis/CommentController.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using RVCoreBoard.MVC.Attributes;
 using RVCoreBoard.MVC.DataContext;
+using RVCoreBoard.MVC.Helpers;
 using RVCoreBoard.MVC.Hubs;
 using RVCoreBoard.MVC.Models;
 using RVCoreBoard.MVC.Services;
@@ -43,6 +44,10 @@ namespace RVCoreBoard.MVC.Apis
         public async Task<IActionResult> CommentDelete(string CNo)
         {
             var comment = await _db.Comments.FirstOrDefaultAsync(c => c.CNo.Equals(int.Parse(CNo)));
+            if (comment == null || User.Identity.GetSid() != comment.UNo)
+            {
+                return Json(new { success = false, responseText = "자신의 댓글만 삭제 할 수 있습니다." });
+            }
 
             _db.Comments.Remove(comment);
             if (_db.SaveChanges() > 0)
@@ -56,9 +61,18 @@ namespace RVCoreBoard.MVC.Apis
         [CustomAuthorize(RoleEnum = UserLevel.Junior | UserLevel.Senior | UserLevel.Manager | UserLevel.Admin)]
         public async Task<IActionResult> CommentModify(Comment comment)
         {
-            comment.Reg_Date = DateTime.Now;
+            if(comment == null || User.Identity.GetSid() != comment.UNo)
+            {
+                return Json(new { success = true, responseText = "자신의 댓글만 수정할 수 있습니다." });
+            }
 
-            _db.Entry(comment).State = EntityState.Modified;
+            //comment.Reg_Date = DateTime.Now;
+            var originComment = await _db.Comments.FirstOrDefaultAsync(p => p.CNo == comment.CNo);
+            if(originComment == null)
+                return Json(new { success = true, responseText = "댓글이 수정되지 않았습니다." });
+
+            originComment.Content = comment.Content;
+            _db.Entry(originComment).State = EntityState.Modified;
             if (await _db.SaveChangesAsync() > 0)
             {
                 return Json(new { success = true, responseText = "댓글이 수정되었습니다." });
@@ -74,6 +88,7 @@ namespace RVCoreBoard.MVC.Apis
             await commentList.GetCommentList(int.Parse(BNo));
 
             return Ok(commentList.Data);
+            //return Json(commentList.Data);
         }
     }
 }
